Start moving to percent-based resizes

+++ zellij-server/src/panes/terminal_pane.rs
@@ -211,46 +211,46 @@ impl Pane for TerminalPane {
     fn pid(&self) -> PaneId {
         PaneId::Terminal(self.pid)
     }
-    fn reduce_height_down(&mut self, count: usize) {
+    fn reduce_height_down(&mut self, count: f64) {
         self.position_and_size.y += count;
         self.position_and_size.rows =
             Dimension::fixed(self.position_and_size.rows.as_usize() - count);
         self.reflow_lines();
     }
-    fn increase_height_down(&mut self, count: usize) {
+    fn increase_height_down(&mut self, count: f64) {
         self.position_and_size.rows =
             Dimension::fixed(self.position_and_size.rows.as_usize() + count);
         self.reflow_lines();
     }
-    fn increase_height_up(&mut self, count: usize) {
+    fn increase_height_up(&mut self, count: f64) {
         self.position_and_size.y -= count;
         self.position_and_size.rows =
             Dimension::fixed(self.position_and_size.rows.as_usize() + count);
         self.reflow_lines();
     }
-    fn reduce_height_up(&mut self, count: usize) {
+    fn reduce_height_up(&mut self, count: f64) {
         self.position_and_size.rows =
             Dimension::fixed(self.position_and_size.rows.as_usize() - count);
         self.reflow_lines();
     }
-    fn reduce_width_right(&mut self, count: usize) {
+    fn reduce_width_right(&mut self, count: f64) {
         self.position_and_size.x += count;
         self.position_and_size.cols =
             Dimension::fixed(self.position_and_size.cols.as_usize() - count);
         self.reflow_lines();
     }
-    fn reduce_width_left(&mut self, count: usize) {
+    fn reduce_width_left(&mut self, count: f64) {
         self.position_and_size.cols =
             Dimension::fixed(self.position_and_size.cols.as_usize() - count);
         self.reflow_lines();
     }
-    fn increase_width_left(&mut self, count: usize) {
+    fn increase_width_left(&mut self, count: f64) {
         self.position_and_size.x -= count;
         self.position_and_size.cols =
             Dimension::fixed(self.position_and_size.cols.as_usize() + count);
         self.reflow_lines();
     }
-    fn increase_width_right(&mut self, count: usize) {
+    fn increase_width_right(&mut self, count: f64) {
         self.position_and_size.cols =
             Dimension::fixed(self.position_and_size.cols.as_usize() + count);
         self.reflow_lines();
